Add functionality to expand text area when clicked on add notes
