+++ src/App.js
@@ -10,6 +10,7 @@ import withNavigation from "./WithNavigation";
 import Logout from "./Logout";
 import Aroute from "./Aroute";
 import UserProfile from "./UserProfile";
+import AddNotes from "./AddNotes";
 
 function App() {
   const HeaderWithNavigation = withNavigation(Header);
@@ -23,6 +24,7 @@ function App() {
             <Route exact path='/register' element={<Register/>}></Route>
             <Route exact path='/login' element={<Login/>}></Route>
             <Route path="/welcome" element={<Aroute><Welcome/></Aroute>} />
+            <Route path="/addnotes" element={<Aroute><AddNotes/></Aroute>} />
             <Route path="/logout" element={<Aroute><Logout/></Aroute>} />
             <Route path="/profile" element={<Aroute><UserProfile/></Aroute>} />
         </Routes>
